src/ai/semantic/requirements.txt
pymilvus==2.4.3
nats-py==2.7.2
protobuf==5.27.0
grpcio==1.63.0
grpcio-tools==1.63.0
python-dotenv==1.0.1
webdriver-manager==4.0.1
selenium==4.21.0
beautifulsoup4==4.12.3
langchain-text-splitters==0.2.0
PyMuPDF==1.24.4
pymupdf4llm==0.0.5
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
sqlalchemy-cockroachdb==2.0.2
youtube-transcript-api==0.6.2
pymupdf4llm==0.0.5
PyMuPDFb==1.24.3
mistune==3.0.2
minio==7.2.7
torch==2.3.0
transformers==4.40.2



